[TODO] keyboard

+++ type.h
@@ -8,5 +8,5 @@ typedef unsigned int  U32;
 typedef long   I64;
 typedef unsigned long U64;
 
-typedef double REAL;
+typedef float REAL;
 #endif
